+++ evm-functional-testing/test/hip-1340/hip1340-sdk-atomic-batch.test.js
@@ -4,83 +4,21 @@ const {hexlify} = require('ethers');
 const hre = require('hardhat');
 const {expect} = require('chai');
 const {
-    AccountCreateTransaction,
-    AccountId,
     AccountInfoQuery,
+    AccountId,
+    AccountCreateTransaction,
     BatchTransaction,
-    EthereumTransaction,
-    EthereumTransactionDataEip7702,
-    PrivateKey,
-    Hbar,
     Client,
-    TransferTransaction,
+    Hbar,
+    PrivateKey,
     ReceiptStatusError,
+    TransferTransaction,
 } = require('@hiero-ledger/sdk');
 const {gas, deploy, getNonces, DelegationTransactionBuilder} = require('./utils/web3');
+const {createEcdsaAliasedAccount, wrapType4ForBatch, expectDelegation} = require('./utils/sdk');
+const sdk = require("@hiero-ledger/sdk");
 
 const SIMPLE_7702_ACCOUNT = '@account-abstraction/contracts/accounts/Simple7702Account';
-
-/**
- * Creates a new ECDSA-aliased Hedera account via the SDK and returns the
- * matching ethers wallet (so callers can sign EVM txs with the same key).
- *
- * @param {import('@hiero-ledger/sdk').Client} client - SDK client used to submit and pay for the AccountCreateTransaction.
- * @param {import('ethers').Provider} provider - ethers provider attached to the returned wallet.
- * @param {Hbar} initialBalance - Starting balance for the new account.
- * @returns {Promise<import('ethers').BaseWallet>} ethers wallet backed by the same ECDSA key as the new Hedera account.
- */
-async function createEcdsaAliasedAccount(client, provider, initialBalance) {
-    const wallet = ethers.Wallet.createRandom(provider);
-    const key = PrivateKey.fromStringECDSA(wallet.privateKey);
-    const receipt = await (
-        await (
-            await new AccountCreateTransaction()
-                .setECDSAKeyWithAlias(key.publicKey)
-                .setInitialBalance(initialBalance)
-                .freezeWith(client)
-                .sign(key)
-        ).execute(client)
-    ).getReceipt(client);
-    expect(receipt.status.toString()).to.equal('SUCCESS');
-    expect(receipt.accountId).to.not.be.null;
-    return wallet;
-}
-
-/**
- * Wraps signed type-4 bytes in an EthereumTransaction batchified under the
- * operator key. fromBytes(...).toBytes() is a byte-equivalent round-trip that
- * acts as a fail-fast validator for the RLP/authorizationList shape.
- *
- * @param {string} rawType4Tx - 0x-prefixed hex of a signed EIP-7702 (type-4) transaction.
- * @param {import('@hiero-ledger/sdk').Client} client - SDK client whose operator key is used as the batch key.
- * @returns {Promise<EthereumTransaction>} An EthereumTransaction frozen and signed by the operator, ready to be added to a BatchTransaction.
- */
-async function wrapType4ForBatch(rawType4Tx, client) {
-    const eip7702Data = EthereumTransactionDataEip7702.fromBytes(
-        Buffer.from(rawType4Tx.slice(2), 'hex')
-    );
-    return new EthereumTransaction()
-        .setEthereumData(eip7702Data.toBytes())
-        .setMaxGasAllowanceHbar(new Hbar(2))
-        .batchify(client, client.operatorPublicKey);
-}
-
-/**
- * Asserts that the given EVM address has its EIP-7702 delegation pointing at
- * the expected smart wallet address.
- *
- * @param {import('@hiero-ledger/sdk').Client} client - SDK client used to issue the AccountInfoQuery.
- * @param {string} evmAddress - 0x-prefixed EVM address of the (aliased) Hedera account whose delegation is being checked.
- * @param {string} expectedDelegation - 0x-prefixed EVM address the account is expected to be delegated to.
- * @returns {Promise<void>}
- */
-async function expectDelegation(client, evmAddress, expectedDelegation) {
-    const accountInfo = await new AccountInfoQuery()
-        .setAccountId(AccountId.fromEvmAddress(0, 0, evmAddress))
-        .execute(client);
-    expect(hexlify(accountInfo.delegationAddress).toLowerCase())
-        .to.equal(expectedDelegation.toLowerCase());
-}
 
 describe('Atomic Batch: EIP-7702 delegation', function () {
     let client, provider, network, smartWalletAddress, sponsor, zeroBalanceAccount;
@@ -99,6 +37,7 @@ describe('Atomic Batch: EIP-7702 delegation', function () {
         sponsor = new ethers.Wallet(hre.network.config.accounts[0], provider);
 
         zeroBalanceAccount = await createEcdsaAliasedAccount(client, provider, new Hbar(0));
+
         const zeroBalanceAccountInfo = await new AccountInfoQuery()
             .setAccountId(AccountId.fromEvmAddress(0, 0, zeroBalanceAccount.address))
             .execute(client);
@@ -139,7 +78,11 @@ describe('Atomic Batch: EIP-7702 delegation', function () {
             ).getReceipt(client);
 
             expect(batchReceipt.status.toString()).to.equal('SUCCESS');
-            await expectDelegation(client, accountA.address, smartWalletAddress);
+            const accountInfo = await new sdk.AccountInfoQuery()
+                .setAccountId(sdk.AccountId.fromEvmAddress(0, 0, accountA.address))
+                .execute(client);
+            expect(hexlify(accountInfo.delegationAddress).toLowerCase())
+                .to.equal(smartWalletAddress.toLowerCase());
         });
 
         it('should commit delegation for pre-existing account used as both from and authority on batch failure', async function () {
@@ -207,7 +150,11 @@ describe('Atomic Batch: EIP-7702 delegation', function () {
             ).getReceipt(client);
 
             expect(batchReceipt.status.toString()).to.equal('SUCCESS');
-            await expectDelegation(client, newAccount.address, smartWalletAddress);
+            const accountInfo = await new sdk.AccountInfoQuery()
+                .setAccountId(sdk.AccountId.fromEvmAddress(0, 0, newAccount.address))
+                .execute(client);
+            expect(hexlify(accountInfo.delegationAddress).toLowerCase())
+                .to.equal(smartWalletAddress.toLowerCase());
         });
     })
